Fix StyleCop warnings in Japanese\TimeParser
